Gen arrays are now double; use new PField table access method.

+++ insts/std/FMINST/FMINST.h
@@ -4,7 +4,7 @@ class FMINST : public Instrument {
 	int nargs, skip, branch;
 	int lenind;
 	float amp, carfreq, carfreqraw, modfreq, modfreqraw, peakdev, spread;
-	float *indexenv, *wavetable, *ampenv;
+	double *indexenv, *wavetable, *ampenv;
 	float indtabs[2], amptabs[2];
 	Ooscili *carosc, *modosc;
